+++ kernel/src/cpu/boot.s
@@ -67,42 +67,3 @@ _start:
 
 .size	_start, . - _start
 .type	_start, function
-
-
-// .section .text.exception_vector_table
-.align 11
-.global exception_vector_table
-exception_vector_table:
-    b svc_handler
-    .align 7
-    b exception_handler
-    .align 7
-    b exception_handler
-    .align 7
-    b exception_handler
-    .align 7
-    b svc_handler
-    .align 7
-    b irq_handler
-    .align 7
-    b exception_handler
-    .align 7
-    b exception_handler
-    .align 7
-    b svc_handler
-    .align 7
-    b irq_handler
-    .align 7
-    b exception_handler
-    .align 7
-    b exception_handler
-    .align 7
-    b exception_handler
-    .align 7
-    b exception_handler
-    .align 7
-    b exception_handler
-    .align 7
-    b exception_handler
-    .align 7
-
