+++ kernel/src/cpu/uart.rs
@@ -1,12 +1,8 @@
 use crate::cpu::registers::Register;
 use crate::cpu::registers::MMIO;
-use crate::os::critical_section::disable_irq;
-use crate::os::critical_section::enable_irq;
-use crate::os::stdio::println;
-use crate::os::stdio::println_now;
-use crate::{println, print};
-use core::arch::asm;
 use crate::os::critical_section;
+use crate::os::stdio::println_now;
+use core::arch::asm;
 
 const BUF_SIZE: usize = 2048;
 
@@ -90,7 +86,7 @@ pub unsafe fn recv_async() -> Option<u8> {
         None
     } else {
         let c = RECV_BUFFER[RECV_START_IDX];
-        
+
         RECV_START_IDX += 1;
         RECV_START_IDX %= BUF_SIZE;
         critical_section::enable_irq();
@@ -98,22 +94,15 @@ pub unsafe fn recv_async() -> Option<u8> {
         // println!("RECV_END_IDX: {}", RECV_END_IDX);
         Some(c)
     }
-    
 }
 
 pub unsafe fn irq_handler() {
-    // if SEND_IDX > 0 {
-    //     println!("Send size: {}", SEND_IDX);
-    // }
-
-    // println_now("IRQ");
-
     for i in 0..SEND_IDX {
         send(SEND_BUFFER[i]);
     }
     SEND_IDX = 0;
-    
-    disable_irq();
+
+    critical_section::disable_irq();
     loop {
         match non_blocking_recv() {
             Some(c) => {
@@ -124,7 +113,6 @@ pub unsafe fn irq_handler() {
             None => break,
         }
     }
-    enable_irq();
+    critical_section::enable_irq();
     MMIO::write(Register::AUX_MU_IER_REG, 0b01);
-    // println!("RECV: {}", RECV_END_IDX);
 }
