+++ kernel/src/main.rs
@@ -4,9 +4,6 @@
 #![feature(btreemap_alloc)]
 #![allow(unused_variables)]
 #![allow(non_camel_case_types)]
-#![allow(unused_assignments)]
-#![allow(unused_mut)]
-#![allow(unused_imports)]
 
 mod cpu;
 mod os;
