+++ kernel/src/os/critical_section.rs
@@ -1,6 +1,5 @@
 use core::arch::asm;
 use core::ptr::{read_volatile, write_volatile};
-use crate::os::stdio::{print_hex_now, println_now};
 
 const CS_COUNTER: *mut u32 = 0x7_5000 as *mut u32;
 
@@ -10,8 +9,6 @@ pub extern "C" fn disable_irq() {
     unsafe {
         let mut counter: u32 = read_volatile(CS_COUNTER);
         counter += 1;
-        // println_now("disable_irq");
-        // print_hex_now(counter);
         write_volatile(CS_COUNTER, counter);
         asm!("msr DAIFSet, 0xf");
     }
@@ -22,8 +19,6 @@ pub extern "C" fn enable_irq() {
     unsafe {
         let mut counter: u32 = read_volatile(CS_COUNTER);
         counter -= 1;
-        // println_now("enable_irq");
-        // print_hex_now(counter);
         write_volatile(CS_COUNTER, counter);
         if counter == 0 {
             asm!("msr DAIFClr, 0xf");
